task 6

+++ 0x00-hello_world/6-size.c
@@ -13,11 +13,11 @@ int main(void)
 	char c;
 	float f;
 
-		printf("Size of a char is %lu.\n", (sizeof(c));
-		printf("Size of an int is %lu.\n", (sizeof(i));
-		printf("Size of a long int is %lu.\n", (sizeof(i));
-		printf("Size of a long long int is %lu.\n", (sizeof(i));
-		printf("Size of a float is %lu.\n", (sizeof(f));
+		printf("Size of a char is %lu.\n", (unsigned long)sizeof(c));
+		printf("Size of an int is %lu.\n", (unsigned long)sizeof(i));
+		printf("Size of a long int is %lu.\n", (unsigned long)sizeof(i));
+		printf("Size of a long long int is %lu.\n", (unsigned long)sizeof(i));
+		printf("Size of a float is %lu.\n", (unsigned long)sizeof(f));
 
 	return (0);
 }
